requirements.txt
pytest==6.1.0
python-telegram-bot==12.8
python-dotenv
requests==2.24.0
django==3.1.2